scripts/requirements.txt
playwright==1.52.0
google-auth==2.38.0
google-api-python-client==2.166.0
